refactor(c3b): Branch module owns element names, linking, teardown, and the WHIP route contract
src/stream/branch.rs is now the only place that knows the per-connection
conventions: whip-sink-{id} / video-queue-{id} / audio-queue-{id} names,
the attach linking + state-sync sequence, the pad-probe teardown dance
(moved verbatim — the hard-won behaviour is untouched), and the loopback
WHIP URL. The route table (startup.rs) and the WHIP Location header both
derive from the same WHIP_SINK_ROUTE template, so the two-files-agreeing-
by-coincidence contract is gone. The bus watch's branch-error containment
check ties to the same naming via is_whip_sink_name.

add_branch/remove_branch become thin calls; link_media's h264/h265
copy-paste arms collapse into one body over a parser table. Pure
naming/path fns are unit-tested (suite now 31 unit tests).

Verified: 31 unit + 10 integration green; --ignored e2e passes in
isolation (17s, clean exit); grep gate confirms no convention definition
outside branch.rs.
