+++ src/stream/mod.rs
@@ -1,7 +1,9 @@
+mod branch;
 mod gst_pipeline;
 mod pipeline;
 mod utils;
 
+pub use branch::{whip_sink_path, WHIP_SINK_ROUTE};
 pub use gst_pipeline::*;
 pub use pipeline::*;
 pub use utils::*;
